rename: modal 폴더 안 파일 join 폴더로 이동
